+++ WBooks/Models/WBooksTabBarController.swift
@@ -18,31 +18,16 @@ class WBooksTabBarController: UITabBarController {
         setupTabs()
     }
     
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        
-        let navigationBar = self.navigationController?.navigationBar
-        
-        navigationBar?.shadowImage = nil
-        navigationBar?.setBackgroundImage(nil, for: .default)
-        navigationBar?.isTranslucent = true
-        
-        navigationBar?.titleTextAttributes = UINavigationBar.appearance().titleTextAttributes
-        navigationBar?.tintColor = UINavigationBar.appearance().tintColor
-    }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        
-        let navigationBar = self.navigationController?.navigationBar
-        
-        navigationBar?.setBackgroundImage(UIImage(), for: .default)
-        navigationBar?.shadowImage = UIImage()
-        navigationBar?.isTranslucent = true
-        
-        let textAttributes = [NSAttributedStringKey.foregroundColor: UIColor.white]
-        navigationBar?.titleTextAttributes = textAttributes
-        navigationBar?.tintColor = UIColor.white
+        setNavigationBarStyle()
+    }
+    
+    
+    override func viewWillDisappear(_ animated: Bool) {
+        super.viewWillDisappear(animated)
+        removeNavigationBarStyle()
     }
     
 }
