+++ WBooks/Recommendation/RecommendationView.xib
@@ -17,14 +17,14 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="ic_add photo" translatesAutoresizingMaskIntoConstraints="NO" id="9aR-9b-2Fs">
-                    <rect key="frame" x="16" y="44" width="130" height="130"/>
+                    <rect key="frame" x="16" y="139" width="130" height="130"/>
                     <constraints>
                         <constraint firstAttribute="width" constant="130" id="i2g-2H-xcn"/>
                         <constraint firstAttribute="height" constant="130" id="s8K-bB-q2J"/>
                     </constraints>
                 </imageView>
                 <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="22" translatesAutoresizingMaskIntoConstraints="NO" id="Ta3-M7-glh">
-                    <rect key="frame" x="16" y="198" width="343" height="313"/>
+                    <rect key="frame" x="16" y="293" width="343" height="313"/>
                     <subviews>
                         <textField opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" textAlignment="natural" minimumFontSize="17" translatesAutoresizingMaskIntoConstraints="NO" id="Hx0-8W-CdO" customClass="SimpleTextField" customModule="WBooks" customModuleProvider="target">
                             <rect key="frame" x="0.0" y="0.0" width="343" height="45"/>
@@ -74,14 +74,14 @@
                     </subviews>
                 </stackView>
                 <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="gKf-xS-M9l">
-                    <rect key="frame" x="16" y="44" width="130" height="130"/>
+                    <rect key="frame" x="16" y="139" width="130" height="130"/>
                     <constraints>
                         <constraint firstAttribute="width" constant="130" id="3qy-9n-2d4"/>
                         <constraint firstAttribute="height" constant="130" id="eMP-td-Quc"/>
                     </constraints>
                 </button>
                 <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="i8p-5t-Yn4" customClass="GradientButton" customModule="WBooks" customModuleProvider="target">
-                    <rect key="frame" x="16" y="543" width="343" height="50"/>
+                    <rect key="frame" x="16" y="638" width="343" height="50"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="50" id="5FM-AX-B5p"/>
                         <constraint firstAttribute="height" relation="greaterThanOrEqual" constant="50" id="oQ7-VK-1eb"/>
@@ -91,19 +91,28 @@
                         <action selector="submit:" destination="iN0-l3-epB" eventType="touchUpInside" id="LgB-oy-AlH"/>
                     </connections>
                 </button>
+                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="bc_nav bar" translatesAutoresizingMaskIntoConstraints="NO" id="TxO-rh-HzS">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="115"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="115" id="bOp-o4-fbq"/>
+                    </constraints>
+                </imageView>
             </subviews>
             <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
             <constraints>
+                <constraint firstItem="gKf-xS-M9l" firstAttribute="top" secondItem="TxO-rh-HzS" secondAttribute="bottom" constant="24" id="0Cn-uF-Qhq"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="i8p-5t-Yn4" secondAttribute="trailing" constant="16" id="3Ak-x0-Lt2"/>
                 <constraint firstItem="Ta3-M7-glh" firstAttribute="top" secondItem="9aR-9b-2Fs" secondAttribute="bottom" constant="24" id="4nW-gP-tBL"/>
                 <constraint firstItem="Ta3-M7-glh" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="74z-Z1-dNf"/>
+                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="TxO-rh-HzS" secondAttribute="trailing" id="Bx6-By-mV5"/>
                 <constraint firstItem="9aR-9b-2Fs" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="SOd-mf-vuK"/>
-                <constraint firstItem="gKf-xS-M9l" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="24" id="Wah-Es-WrM"/>
+                <constraint firstItem="TxO-rh-HzS" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="ZKg-kz-mGo"/>
                 <constraint firstItem="i8p-5t-Yn4" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="ZfQ-6l-rjL"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="Ta3-M7-glh" secondAttribute="trailing" constant="16" id="ctR-NY-dNc"/>
                 <constraint firstItem="gKf-xS-M9l" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="eNe-AT-QTG"/>
+                <constraint firstItem="9aR-9b-2Fs" firstAttribute="top" secondItem="TxO-rh-HzS" secondAttribute="bottom" constant="24" id="lfi-gk-1oy"/>
                 <constraint firstItem="i8p-5t-Yn4" firstAttribute="top" secondItem="Ta3-M7-glh" secondAttribute="bottom" constant="32" id="rKt-oD-5kk"/>
-                <constraint firstItem="9aR-9b-2Fs" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="24" id="zhk-w0-Osb"/>
+                <constraint firstItem="TxO-rh-HzS" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="w4m-o1-NG4"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
             <connections>
@@ -120,6 +129,7 @@
         </view>
     </objects>
     <resources>
+        <image name="bc_nav bar" width="375" height="102"/>
         <image name="ic_add photo" width="96" height="96"/>
     </resources>
 </document>
