+++ WBooks/Rentals/RentalsView.xib
@@ -41,18 +41,27 @@
                     </subviews>
                 </stackView>
                 <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="28" sectionFooterHeight="28" translatesAutoresizingMaskIntoConstraints="NO" id="3A8-vJ-shv">
-                    <rect key="frame" x="16" y="20" width="343" height="460.5"/>
+                    <rect key="frame" x="16" y="115" width="343" height="365"/>
                     <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                 </tableView>
+                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="bc_nav bar" translatesAutoresizingMaskIntoConstraints="NO" id="VSA-Mt-sVd">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="115"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="115" id="HqT-fm-rT0"/>
+                    </constraints>
+                </imageView>
             </subviews>
             <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
             <constraints>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="3A8-vJ-shv" secondAttribute="trailing" constant="16" id="9mZ-Rs-3dB"/>
+                <constraint firstItem="VSA-Mt-sVd" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="LiN-h9-PZN"/>
+                <constraint firstItem="VSA-Mt-sVd" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="QFV-aD-3LU"/>
+                <constraint firstItem="3A8-vJ-shv" firstAttribute="top" secondItem="VSA-Mt-sVd" secondAttribute="bottom" id="Qtf-Za-K1u"/>
                 <constraint firstItem="3A8-vJ-shv" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="gc5-ec-Yxj"/>
                 <constraint firstAttribute="trailing" secondItem="CW3-iE-V7F" secondAttribute="trailing" constant="16" id="hJZ-0r-upQ"/>
                 <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="CW3-iE-V7F" secondAttribute="bottom" id="ibT-Ao-L0h"/>
+                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="VSA-Mt-sVd" secondAttribute="trailing" id="kMu-am-ekY"/>
                 <constraint firstItem="CW3-iE-V7F" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="16" id="ke1-jI-ZjP"/>
-                <constraint firstItem="3A8-vJ-shv" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" id="q06-Eq-Q65"/>
                 <constraint firstItem="CW3-iE-V7F" firstAttribute="top" secondItem="3A8-vJ-shv" secondAttribute="bottom" constant="16" id="zTP-Gb-WO5"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
@@ -63,4 +72,7 @@
             <point key="canvasLocation" x="26.5" y="52.5"/>
         </view>
     </objects>
+    <resources>
+        <image name="bc_nav bar" width="375" height="102"/>
+    </resources>
 </document>
