+++ WBooks/Wishlist/WishlistView.xib
@@ -16,21 +16,23 @@
             <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Wishlist" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="u02-rd-O9o">
-                    <rect key="frame" x="157" y="323" width="60" height="21"/>
-                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
+                <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="bc_nav bar" translatesAutoresizingMaskIntoConstraints="NO" id="l82-fP-Ppr">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="115"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="115" id="gLc-TK-fsb"/>
+                    </constraints>
+                </imageView>
             </subviews>
             <color key="backgroundColor" red="1" green="1" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
             <constraints>
-                <constraint firstItem="u02-rd-O9o" firstAttribute="top" secondItem="vUN-kp-3ea" secondAttribute="top" constant="303" id="2dO-fA-R9N"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="u02-rd-O9o" secondAttribute="bottom" constant="323" id="X32-2M-puh"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="u02-rd-O9o" secondAttribute="trailing" constant="158" id="f1v-Hk-OFW"/>
-                <constraint firstItem="u02-rd-O9o" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="157" id="hsW-sV-xTL"/>
+                <constraint firstItem="l82-fP-Ppr" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="A2t-6S-G0g"/>
+                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="l82-fP-Ppr" secondAttribute="trailing" id="UEC-GE-uXq"/>
+                <constraint firstItem="l82-fP-Ppr" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="yM5-H8-J2h"/>
             </constraints>
             <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
         </view>
     </objects>
+    <resources>
+        <image name="bc_nav bar" width="375" height="102"/>
+    </resources>
 </document>
